Capacitor output to file

+++ barna.capacitor/src/improvementTest/groovy/barna/flux/capacitor/improvementtest/FluxCapacitorRunImprovementTest.groovy
@@ -1,7 +1,12 @@
 package barna.flux.capacitor.improvementtest
 
 import barna.commons.Execute
+import barna.flux.capacitor.reconstruction.FluxCapacitorSettings
+import barna.io.FileHelper
 import org.rosuda.JRI.Rengine
+
+import java.util.concurrent.Executors
+
 import org.junit.*
 
 /**
@@ -43,8 +48,8 @@ class FluxCapacitorRunImprovementTest {
     File currentTestDirectory = null
     @Before
     public void setUpTest(){
-//        currentTestDirectory = FileHelper.createTempDir("FluxCapacitorImprovement", "", null)
-        currentTestDirectory = new File("FluxCapacitorImprovement", new File("/tmp"))
+        currentTestDirectory = FileHelper.createTempDir("FluxCapacitorImprovement", "", null)
+//        currentTestDirectory = new File("FluxCapacitorImprovement", new File("/tmp"))
     }
 
     @After
@@ -56,11 +61,12 @@ class FluxCapacitorRunImprovementTest {
 
     @Test
     public void testDeconvolution() {
-        /*println "======Running deconvolution======"
+        println "======Running deconvolution======"
         def pool = Executors.newFixedThreadPool(THREADS)
         println "Setting up ${SAMPLES/THREADS} pools with $THREADS slots"
         def bams = new File(FluxCapacitorRunner.testData['bam'])
         def c = 0
+        def stdout = []
         bams.eachFileMatch(~/^.*\.bam$/) {
             def name = it.toString().split("/").last().replace(".bam","")
             def fileDir = FileHelper.createTempDir(name, "",currentTestDirectory)
@@ -72,9 +78,10 @@ class FluxCapacitorRunImprovementTest {
             ])
 
             def i = c++%THREADS
+            stdout[i] = ""
             def t = new Thread() {
                 public void run() {
-                    barna.flux.capacitor.improvementtest.FluxCapacitorRunner.runCapacitor(fileDir,parFile)
+                    stdout [i] = barna.flux.capacitor.improvementtest.FluxCapacitorRunner.runCapacitor(fileDir,parFile)
                 }
             }
 
@@ -85,6 +92,10 @@ class FluxCapacitorRunImprovementTest {
                 print "Executing threads in pool ${c/THREADS}..."
                 while (!pool.isTerminated()) {}
                 println "done"
+                stdout.eachWithIndex {
+                    File out = new File("stderr." + c/THREADS + "." + i + ".gtf",currentTestDirectory)
+                    out.append(it)
+                }
                 pool = Executors.newFixedThreadPool(THREADS)
             }
         }
@@ -116,14 +127,14 @@ class FluxCapacitorRunImprovementTest {
                 j++
             }
             i++
-        }*/
+        }
         def table = new File(currentTestDirectory.absolutePath+"/transcriptsRPKM.txt")
-        /*if (table.exists())
+        if (table.exists())
             table.delete()
         file.each {
             table.append(it)
             table.append(barna.commons.system.OSChecker.NEW_LINE)
-        }*/
+        }
         println "======Building correlation matrix of samples======"
         def corMat = new File(currentTestDirectory.absolutePath+"/corMat.txt")
         def pb = new ProcessBuilder()
